🚀 feat(get-user-metrics-usecase): add GetUserMetricsUseCase to calculate user metrics based on meals
🚧 test(get-user-metrics-usecase.spec): add unit test for GetUserMetricsUseCase

The GetUserMetricsUseCase is a new use case that calculates user metrics based on their meals. It takes a userId as input and returns the following metrics:
- totalMealsRecorded: the total number of meals recorded by the user
- totalMealsInDiet: the total number of meals that are in the user's diet
- totalMealsOutsideDiet: the total number of meals that are outside the user's diet
- bestSequenceMealsInDiet: the longest sequence of consecutive meals in the user's diet

The use case relies on the MealsRepository to fetch the meals for the user. The GetUserMetricsUseCase is tested with a unit test in the get-user-metrics-usecase.spec file.